Iris - Implicit conversion fixes.
#rb mattias.hornlund
#rnx

[CL 29310644 by peter engstrom in ue5-main branch]

+++ Engine/Source/Runtime/Experimental/Iris/Core/Private/Iris/ReplicationSystem/Conditionals/ReplicationConditionals.cpp
@@ -65,7 +65,7 @@ bool FReplicationConditionals::SetConditionConnectionFilter(FInternalNetRefIndex
 	if (ObjectInfo->AutonomousConnectionId != AutonomousConnectionId)
 	{
 		const uint32 ConnIdForBaselineInvalidation = (bEnable ? ConnectionId : ObjectInfo->AutonomousConnectionId);
-		ObjectInfo->AutonomousConnectionId = AutonomousConnectionId;
+		ObjectInfo->AutonomousConnectionId = uint16(AutonomousConnectionId);
 
 		BaselineInvalidationTracker->InvalidateBaselines(ObjectIndex, ConnIdForBaselineInvalidation);
 	}
